refactor: extract issue report state from AgentWorkspace
