+++ src/components/agent/AgentWorkspace.tsx
@@ -102,6 +102,7 @@ import {
 } from "../../lib/agent-chat-gallery";
 import { attachScrollThumbFade } from "../../lib/scroll-thumb-fade";
 import type { AgentWorkspaceProps } from "./agent-workspace-types";
+import { createIssueReportStateActions } from "./issue-report-state-actions";
 import { createAttachmentImportActions } from "./attachment-import-actions";
 import { createQueuedFollowUpRenderers } from "./queued-follow-up-renderers";
 import { useAgentSessionLoading } from "./use-agent-session-loading";
@@ -241,7 +242,6 @@ import {
 } from "./composer/follow-up-queue";
 
 import {
-  persistReviewableIssueReports,
   persistedReviewableIssueReports,
   rememberComposerDraft,
   NEW_SESSION_RECOVERY_QUEUE_KEY,
@@ -704,83 +704,25 @@ export function AgentWorkspace({
     prompt: string;
   } | null>(null);
 
-  function setReviewableIssueReport(sessionId: string, report: PendingIssueReport | null) {
-    const next = { ...reviewableIssueReportsRef.current };
-    if (report) {
-      next[sessionId] = report;
-    } else {
-      delete next[sessionId];
-    }
-    reviewableIssueReportsRef.current = next;
-    persistReviewableIssueReports(next);
-    setReviewableIssueReports(next);
-  }
-
-  function setIssueReportDiagnosisRefreshing(sessionId: string, refreshing: boolean) {
-    const next = new Set(diagnosisRefreshIssueReportSessionIdsRef.current);
-    if (refreshing) {
-      next.add(sessionId);
-    } else {
-      next.delete(sessionId);
-    }
-    diagnosisRefreshIssueReportSessionIdsRef.current = next;
-    setDiagnosisRefreshIssueReportSessionIds(next);
-  }
-
-  function queueIssueReportDiagnosisRefresh(sessionId: string, delayMs = 300) {
-    setIssueReportDiagnosisRefreshing(sessionId, true);
-    let refresh: Promise<void>;
-    refresh = new Promise<void>((resolve) => {
-      window.setTimeout(() => {
-        void refreshHermesSession(sessionId).finally(resolve);
-      }, delayMs);
-    }).finally(() => {
-      if (issueReportDiagnosisRefreshesRef.current.get(sessionId) === refresh) {
-        issueReportDiagnosisRefreshesRef.current.delete(sessionId);
-        setIssueReportDiagnosisRefreshing(sessionId, false);
-      }
-    });
-    issueReportDiagnosisRefreshesRef.current.set(sessionId, refresh);
-    return refresh;
-  }
-
-  function waitForIssueReportDiagnosisRefresh(sessionId: string) {
-    if (!diagnosisRefreshIssueReportSessionIdsRef.current.has(sessionId)) {
-      return Promise.resolve();
-    }
-    return (
-      issueReportDiagnosisRefreshesRef.current.get(sessionId) ??
-      queueIssueReportDiagnosisRefresh(sessionId)
-    );
-  }
-
-  function promotePendingIssueReportToReview(
-    sessionId: string,
-    options: { queueDiagnosisRefresh: boolean },
-  ) {
-    const issueReport = pendingIssueReportsRef.current.get(sessionId);
-    if (!issueReport) return false;
-    pendingIssueReportsRef.current.delete(sessionId);
-    deferredFailedIssueReportDeliverySessionIdsRef.current.delete(sessionId);
-    setReviewableIssueReport(sessionId, issueReport);
-    if (options.queueDiagnosisRefresh) {
-      queueIssueReportDiagnosisRefresh(sessionId);
-    } else {
-      setIssueReportDiagnosisRefreshing(sessionId, false);
-    }
-    return true;
-  }
-
-  function setIssueReportSubmitting(sessionId: string, submitting: boolean) {
-    const next = new Set(submittingIssueReportSessionIdsRef.current);
-    if (submitting) {
-      next.add(sessionId);
-    } else {
-      next.delete(sessionId);
-    }
-    submittingIssueReportSessionIdsRef.current = next;
-    setSubmittingIssueReportSessionIds(next);
-  }
+  const {
+    setReviewableIssueReport,
+    setIssueReportDiagnosisRefreshing,
+    queueIssueReportDiagnosisRefresh,
+    waitForIssueReportDiagnosisRefresh,
+    promotePendingIssueReportToReview,
+    setIssueReportSubmitting,
+  } = createIssueReportStateActions({
+    deferredFailedIssueReportDeliverySessionIdsRef,
+    diagnosisRefreshIssueReportSessionIdsRef,
+    issueReportDiagnosisRefreshesRef,
+    pendingIssueReportsRef,
+    refreshHermesSession,
+    reviewableIssueReportsRef,
+    setDiagnosisRefreshIssueReportSessionIds,
+    setReviewableIssueReports,
+    setSubmittingIssueReportSessionIds,
+    submittingIssueReportSessionIdsRef,
+  });
 
   useIssueReportEvents({
     deferredFailedIssueReportDeliverySessionIdsRef,
